// facebook-api/src/facebook/models/FacebookInterfaces.ts
export interface GPTads {
  ad_creative_bodies: string;
  age_country_gender_reach_breakdown: any[]; // Consider specifying further if possible
  eu_total_reach: number;
  publisher_platforms: string[];
  target_ages: string[];
  target_gender: string;
  target_locations: any[]; // Consider specifying further if possible
  creative_src: string;
  full_ad_src: string;
  landing_page_src: string;
  ad_delivery_start_time: string;
}
// ...
export interface ChatCompletionResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  system_fingerprint: string;
  choices: Choice[];
  usage: Usage;
  error?: {
    message: string;
    type: string;
    param: string;
    code: string;
  };
}

export interface Choice {
  index: number;
  message: MessageResponse;
  logprobs: null; // Assuming logprobs is always null based on the provided structure. Adjust accordingly if it varies.
  finish_reason: string;
}

export interface MessageResponse {
  content: string;
}

export interface Usage {
  prompt_tokens: number;
  completion_tokens: number;
  total_tokens: number;
}

export interface InputMessageType {
  type: string;
  text?: string;
  image_url?: {
    url: string;
  };
}

export interface TextContent {
  type: 'text';
  text: string;
}

export interface ImageContent {
  type: 'image_url';
  image_url: {
    url: string;
  };
}

export type Content = TextContent | ImageContent;

export interface MessagePayload {
  role: string;
  content: Content[]; // Now content is an array of Content
}

export interface GPTPayload {
  model: string;
  messages: MessagePayload[];
  max_tokens: number;
  temperature: number;
}